pertemuan 6

+++ praktikum/app/Http/Controllers/Studentcontroller.php
@@ -15,67 +15,92 @@ class Studentcontroller extends Controller
         //Menggunakan model Student untuk select data
         $students = Student::all();
         
+        //jika data kosong maka kirim status code 204
+        if($students->isEmpty()){
+            $data = [
+                'message' => 'Resource Is Empty'
+        ];
+        //menembalikan data dalam bentuk json
+            return response()->json($data, 200);
+     }
+
         $data = [
             'message' => 'Get all students',
             'data' => $students
         ];
-        //menembalikan data dalam bentuk json
+
+        //Mengembalikan data dalam bentuk json
         return response()->json($data, 200);
+    
     }
 
-    /**
-     * Store a newly created resource in storage.
-     */
     public function store(Request $request)
     {
+        // Validasi data request
+        $request->validate([
+            'nama' => "required",
+            'nim' => "required",
+            'email' => "required|email",
+            'jurusan' => "required"
+        ]);
+
        $input = [
-        'nama' => $request->nama,
-        'nim' => $request->nim,
-        'email' => $request->email,
-        'jurusan' => $request->jurusan
+            'nama' => $request->nama,
+            'nim' => $request->nim,
+            'email' => $request->email,
+            'jurusan' => $request->jurusan
        ];
 
         $student = Student::create($input);
 
         $data = [
             'message' => 'Student is created succesfully',
-            'data' => $student,
+            'data' => $student
         ];
         
         return response()->json($data, 201);
     }
 
-    /**
-     * Update the specified resource in storage.
-     */
     public function update(Request $request , $id)
     {
         //
         $student = Student::find($id);
 
-        $input = [
-            'nama' => $request->nama,
-            'nim' => $request->nim,
-            'email' => $request->email,
-            'jurusan' => $request->jurusan
-        ];
+        if (!$student) {
+            $data = [
+                "messaga" => 'Data not found'
+            ];
 
-        $student->update($input);
+            return response()->json($data, 404);
+        }
+
+        $student->update([
+            'nama' => $request->nama ?? $student->nama,
+            'nim' => $request->nim ?? $student->nim,
+            'email' => $request->email ?? $student->email,
+            'jurusan' => $request->jurusan ?? $student->jurusan
+       ]);
 
         $data = [
             'message' => 'Student is updated succesfully',
             'data' => $student
         ];
-            return response()->json($data, 404);
+            return response()->json($data, 200);
     }
 
-    /**
-     * Remove the specified resource from storage.
-     */
+
     public function destroy($id)
     {
         //
-        $students = Student::find($id);
+        $student = Student::find($id);
+
+        if(!$student){
+            $data = [
+                'message' => 'Student not dound'
+            ];
+
+            return response()->json($data, 404);
+        }
 
         $student->delete();
 
@@ -84,6 +109,28 @@ class Studentcontroller extends Controller
             'data' => $student
         ];
 
-            return response()->json($response, 404);
+            return response()->json($response, 200);
+}
+
+public function show ($id)
+{
+    $student = Student::find($id);
+
+    if (!student){
+        $data = [
+
+            "message" => " Data not found"
+        ];
+        
+
+        return response()->json($data, 404);
+    }
+
+    $data - [
+        "message" => "Show detail resource",
+        "data" => $student
+    ];
+
+    return response()->json($data, 200);
     }
 }
